De-active cross-version test retry
This no longer works because the daemon fixtures tries to instantiate
the daemon home service, which may fail with an older Tooling API
on the test classpath.

It could be fixed by changing the daemon fixture such that it works
without the service. Or we remove the retry logic in favour of the
test retry plugin that we now use.